feat: show the Login and Logout button if the user is logged out / logged in

+++ youtube-client-app/src/app/auth/services/auth.service.ts
@@ -1,25 +1,43 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
+import { BehaviorSubject, Observable, Observer, Subject } from 'rxjs';
 import { SearchDataService } from 'src/app/youtube/services/search-data.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class AuthService {
+  loginSubject = new BehaviorSubject<boolean>(false)
   constructor(
     private router: Router,
     private searchDataService: SearchDataService,
-  ) {}
+  ) {
+    if (localStorage.getItem('login')) {
+      this.loginSubject.next(true)
+    }
+    this.loginSubject.subscribe({
+      next: (val) => {
+        console.log(val)
+        if (val) {
+          localStorage.setItem('login', 'test');
+          this.router.navigate(['']);
+        } else {
+          localStorage.removeItem('login');
+          this.router.navigate(['login']);
+          this.searchDataService.deleteResultData();
+        }
+      },
+      error: (err) => console.log({"err": err}),
+      complete: () => console.log('subject-complete')
+    })
+  }
 
   login() {
-    localStorage.setItem('login', 'test');
-    this.router.navigate(['']);
+    this.loginSubject.next(true)
   }
 
   logout() {
-    localStorage.removeItem('login');
-    this.router.navigate(['login']);
-    this.searchDataService.deleteResultData();
+    this.loginSubject.next(false)    
   }
 
   static isAuthCheck(): Promise<boolean> {
